Drop CommonWeb dep on AppStrings
Inject strings from a generic Bundle.
