+++ Packages/App/Tests/CommonWebTests/Resources/en.lproj/Localizable.strings
@@ -1,0 +1,1 @@
+"web_uploader.success" = "Upload complete!";
